CKVS -> CKVSImpl
